Use pytest_addoption to configure tests
