+++ tox.ini
@@ -6,9 +6,9 @@ extras = dev
 deps =
     pyyaml
 commands =
-    pytest
+    pytest tests/ --ros-root /opt/ros/melodic --python-version 2.7
 setenv =
-    PYTHONPATH = /opt/ros/noetic/lib/python2.7/dist-packages
+    PYTHONPATH = /opt/ros/melodic/lib/python2.7/dist-packages
 
 [testenv:py38]
 skip_install = true
